requirements.txt
fastapi
uvicorn
gradio
pytesseract
Pillow
requests
pandas
openpyxl
python-dotenv